Camera changes and more code structure
Added a camera based on pos + target, using basic camera matrix calculation to work out directions based on UVs. Created a ray structure that can be initialised via this camera setup, put basics in place to start iterating rays and bouncing them. Will probably work on the assumption that if a ray doesn't hit a sphere then it'll hit a wall, i.e. it'll never hit a wall first.